requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
networkx==3.2.1
pyvis==0.3.2
reportlab==4.1.0
torch==2.2.2
transformers==4.39.3
sentence-transformers==2.6.1
umap-learn==0.5.5
hdbscan==0.8.33

